+++ src/main/java/controledenotas/business/entity/ProfessorBC.java
@@ -9,7 +9,7 @@ import br.gov.frameworkdemoiselle.transaction.Transactional;
 import br.gov.frameworkdemoiselle.template.DelegateCrud;
 
 @BusinessController
-public class ProfessorBC extends DelegateCrud<Professor, Integer, ProfessorDAO> {
+public class ProfessorBC extends DelegateCrud<Professor, Long, ProfessorDAO> {
 
 	private static final long serialVersionUID = 1L;
